Separate JS modules into subdirectory
