+++ compiler/core/src/main.bs.js
@@ -1,19 +1,20 @@
 // Generated by BUCKLESCRIPT VERSION 2.1.0, PLEASE EDIT WITH CARE
 'use strict';
 
-var Fs                              = require("fs");
-var List                            = require("bs-platform/lib/js/list.js");
-var Glob                            = require("glob");
-var Path                            = require("path");
-var $$Array                         = require("bs-platform/lib/js/array.js");
-var Process                         = require("process");
-var FsExtra                         = require("fs-extra");
-var Caml_array                      = require("bs-platform/lib/js/caml_array.js");
-var Color$LonaCompilerCore          = require("./color.bs.js");
-var JavaScript$LonaCompilerCore     = require("./javaScript.bs.js");
-var SwiftColor$LonaCompilerCore     = require("./swift/swiftColor.bs.js");
-var SwiftRender$LonaCompilerCore    = require("./swift/swiftRender.bs.js");
-var SwiftComponent$LonaCompilerCore = require("./swift/swiftComponent.bs.js");
+var Fs                                   = require("fs");
+var List                                 = require("bs-platform/lib/js/list.js");
+var Glob                                 = require("glob");
+var Path                                 = require("path");
+var $$Array                              = require("bs-platform/lib/js/array.js");
+var Process                              = require("process");
+var FsExtra                              = require("fs-extra");
+var Caml_array                           = require("bs-platform/lib/js/caml_array.js");
+var Color$LonaCompilerCore               = require("./color.bs.js");
+var SwiftColor$LonaCompilerCore          = require("./swift/swiftColor.bs.js");
+var SwiftRender$LonaCompilerCore         = require("./swift/swiftRender.bs.js");
+var SwiftComponent$LonaCompilerCore      = require("./swift/swiftComponent.bs.js");
+var JavaScriptRender$LonaCompilerCore    = require("./javaScript/javaScriptRender.bs.js");
+var JavaScriptComponent$LonaCompilerCore = require("./javaScript/javaScriptComponent.bs.js");
 
 function exit(message) {
   console.log(message);
@@ -99,7 +100,7 @@ function convertComponent(filename) {
       return (process.exit());
     }
   } else {
-    return JavaScript$LonaCompilerCore.Render[/* toString */2](JavaScript$LonaCompilerCore.Component[/* generate */0](name, parsed));
+    return JavaScriptRender$LonaCompilerCore.toString(JavaScriptComponent$LonaCompilerCore.generate(name, parsed));
   }
 }
